requirements.txt
aiohttp>=3.9
httpx>=0.27
pyyaml>=6.0
